+++ drv/i2c-api/src/lib.rs
@@ -197,7 +197,7 @@ impl Marshal<[u8; 4]> for I2cMessage {
         [
             self.0,
             self.1 as u8,
-            self.2 .0 as u8,
+            self.2 .0,
             match self.3 {
                 Some((mux, seg)) => {
                     0b1000_0000 | ((mux as u8) << 4) | (seg as u8)
